move logic to dedicated hook
